Moved usage to test file
